add command line profile
