+++ CMakeLists.txt
@@ -39,7 +39,7 @@ option(MICM_ENABLE_OPENACC "Build with OpenACC Support" OFF)
 option(MICM_ENABLE_LLVM "Build with LLVM support for JIT-compiling" OFF)
 option(MICM_ENABLE_TESTS "Build the tests" ON)
 option(MICM_ENABLE_EXAMPLES "Build the examples" ON)
-option(MICM_ENABLE_PROFILE "Add the instrumentation" OFF)
+option(MICM_ENABLE_PROFILE "Profile MICM Solver" ON)
 set(MICM_DEFAULT_VECTOR_MATRIX_SIZE "4" CACHE STRING "Default size for vectorizable matrix types")
 
 include(CMakeDependentOption)
@@ -108,17 +108,10 @@ if (${CMAKE_HOST_SYSTEM_NAME} MATCHES "Linux" AND "${CMAKE_CXX_COMPILER_ID}" STR
 endif()
 
 ################################################################################
-# Command-line driver examples
+# Command-line driver examples / Command-line profiler
 
-if(PROJECT_IS_TOP_LEVEL AND MICM_ENABLE_EXAMPLES AND MICM_ENABLE_JSON)
+if(PROJECT_IS_TOP_LEVEL AND MICM_ENABLE_JSON AND (MICM_ENABLE_EXAMPLES OR MICM_ENABLE_PROFILE))
   add_subdirectory(examples)
-endif()
-
-################################################################################
-# Command-line profiler
-
-if(PROJECT_IS_TOP_LEVEL AND MICM_ENABLE_PROFILE AND MICM_ENABLE_JSON)
-  add_subdirectory(profiling)
 endif()
 
 ################################################################################
